+++ examples/with-thirdpartyemailpassword-express/index.js
@@ -98,12 +98,6 @@ supertokens.init({
                     clientId: "1060725074195-kmeum4crr01uirfl2op9kd5acmi9jutn.apps.googleusercontent.com",
                     clientSecret: "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",
                 }),
-
-                // we have commented the below because our app domain (ThirdPartyEmailPassword.demo.supertokens.io) is not approved by Facebook since it's only a demo app.
-                // ThirdPartyEmailPassword.Facebook({
-                //     clientSecret: process.env.FACEBOOK_CLIENT_SECRET,
-                //     clientId: process.env.FACEBOOK_CLIENT_ID
-                // })
             ],
         }),
         Session.init(),