+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 .
 
+## [Unreleased]
+
+### Added
+
+-   Sign in with Discord, Google workspaces
+-   If `getProfileInfo` throws an error, then it is sent as a FIELD_ERROR to the frontend
+
 ## [8.1.0] - 2021-10-29
 
 ### Added